refactor(db): extract checkpoint recovery and handoff
