+++ crates/laminar-db/src/checkpoint_coordinator/mod.rs
@@ -6,6 +6,9 @@
 mod follower_completion;
 #[cfg(feature = "cluster")]
 mod follower_prepare;
+#[cfg(feature = "cluster")]
+mod handoff;
+mod recovery;
 mod retention;
 pub(crate) mod sink_epoch_admission;
 #[cfg(feature = "cluster")]
@@ -17,7 +20,7 @@ use std::sync::Arc;
 use std::time::{Duration, Instant};
 
 use bytes::Bytes;
-use futures::{stream::FuturesOrdered, stream::FuturesUnordered, StreamExt, TryStreamExt};
+use futures::{stream::FuturesOrdered, stream::FuturesUnordered, StreamExt};
 use laminar_connectors::checkpoint::SourceCheckpoint;
 use laminar_connectors::connector::{
     CoordinatedCommitBatch, CoordinatedCommitCursor, CoordinatedCommitNamespace,
@@ -44,10 +47,6 @@ use sha2::{Digest, Sha256};
 use tracing::warn;
 
 use crate::error::DbError;
-#[cfg(feature = "cluster")]
-use crate::recovery_manager::{
-    load_verified_state_frames, RecoveredStateFrame, VerifiedStateFramePlan,
-};
 use retention::{run_gc_worker, GcRequest};
 
 const MAX_SINK_PHASE_ONE_CONCURRENCY: usize = 8;
@@ -312,11 +311,6 @@ async fn checkpoint_kill_gate(
         tokio::time::sleep(Duration::from_millis(5)).await;
     }
     let _ = std::fs::remove_file(ready_file);
-}
-
-#[cfg(feature = "cluster")]
-fn handoff_error(message: impl Into<String>) -> DbError {
-    DbError::Checkpoint(format!("[LDB-6050] {}", message.into()))
 }
 
 #[derive(Debug, Clone)]
@@ -703,40 +697,6 @@ impl<'a> PrepareQuorum<'a> {
 pub(crate) struct RegisteredSink {
     name: String,
     handle: crate::sink_task::SinkTaskHandle,
-}
-
-#[cfg(feature = "cluster")]
-fn recovery_sink_fence(
-    checkpoint_proof: Option<&LeaderProof>,
-    continuation_proof: Option<&LeaderProof>,
-) -> Result<Option<u64>, DbError> {
-    let Some(continuation_proof) = continuation_proof else {
-        return Ok(None);
-    };
-    if !continuation_proof.is_canonical() {
-        return Err(DbError::Checkpoint(
-            "cluster recovery continuation has a non-canonical leader proof".into(),
-        ));
-    }
-    let checkpoint_proof = checkpoint_proof
-        .filter(|proof| proof.is_canonical())
-        .ok_or_else(|| {
-            DbError::Checkpoint(
-                "cluster recovery checkpoint has no canonical publishing leader proof".into(),
-            )
-        })?;
-    if continuation_proof.fencing_token < checkpoint_proof.fencing_token {
-        return Err(DbError::Checkpoint(format!(
-            "cluster recovery leader fencing token {} regressed below checkpoint token {}",
-            continuation_proof.fencing_token, checkpoint_proof.fencing_token
-        )));
-    }
-
-    // RECOVERY: the immutable checkpoint outcome binds the exact Delta publication batch to the
-    // leader term that committed it. A successor is the designated reconciler, but changing the
-    // batch token would make an already-published exact cursor look conflicting instead of
-    // idempotent after failover.
-    Ok(Some(checkpoint_proof.fencing_token))
 }
 
 struct PackedCheckpoint {
@@ -1362,12 +1322,6 @@ impl CheckpointCoordinator {
         &self.config
     }
 
-    #[cfg(feature = "cluster")]
-    pub(crate) fn set_recovery_graph_payload_limit(&mut self, bytes: usize) {
-        debug_assert_ne!(bytes, 0);
-        self.recovery_graph_payload_limit = bytes;
-    }
-
     #[must_use]
     pub(crate) fn epoch_allocator(&self) -> Arc<EpochAllocator> {
         Arc::clone(&self.allocator)
@@ -1376,688 +1330,6 @@ impl CheckpointCoordinator {
     #[must_use]
     pub fn store(&self) -> &dyn CheckpointStore {
         self.store.as_ref()
-    }
-
-    #[cfg(feature = "cluster")]
-    pub(crate) async fn load_handoff_state_frames(
-        &self,
-        pinned: &CommittedCheckpointRef,
-        predecessor: &laminar_core::checkpoint::CheckpointAssignmentFence,
-        predecessor_owners: &[laminar_core::state::NodeId],
-        acquired_vnodes: &[u32],
-        include_whole: bool,
-        max_payload_bytes: usize,
-        deadline: tokio::time::Instant,
-    ) -> Result<Vec<RecoveredStateFrame>, DbError> {
-        if tokio::time::Instant::now() >= deadline {
-            return Err(handoff_error("vnode handoff read timed out"));
-        }
-        let owner_ids = predecessor_owners
-            .iter()
-            .map(|owner| owner.0)
-            .collect::<Vec<_>>();
-        let key_group_count = self.store.key_group_count();
-        if !predecessor.is_canonical()
-            || predecessor.vnode_count != u32::from(key_group_count)
-            || owner_ids.len() != usize::from(key_group_count.get())
-            || !predecessor.matches_owner_map(&owner_ids)
-        {
-            return Err(handoff_error(
-                "predecessor fence does not match the exact vnode owner map",
-            ));
-        }
-        if acquired_vnodes.is_empty()
-            || acquired_vnodes.windows(2).any(|pair| pair[0] >= pair[1])
-            || acquired_vnodes
-                .iter()
-                .any(|vnode| *vnode >= predecessor.vnode_count)
-        {
-            return Err(handoff_error(
-                "acquired vnode roster must be nonempty, canonical, and in range",
-            ));
-        }
-
-        let decision_store = self.decision_store.as_ref().ok_or_else(|| {
-            handoff_error("vnode handoff requires a durable checkpoint decision store")
-        })?;
-        let committed =
-            tokio::time::timeout_at(deadline, decision_store.load_committed_checkpoint(pinned))
-                .await
-                .map_err(|_| handoff_error("committed handoff checkpoint read timed out"))?
-                .map_err(|error| {
-                    handoff_error(format!("committed handoff checkpoint read failed: {error}"))
-                })?;
-        let pipeline_identity = self.expected_pipeline_identity()?;
-        let deployment_id = self.expected_deployment_id()?.to_owned();
-        if committed.pipeline_identity != pipeline_identity {
-            return Err(handoff_error(
-                "handoff checkpoint pipeline identity does not match the active pipeline",
-            ));
-        }
-        if committed.deployment_id != deployment_id {
-            return Err(handoff_error(
-                "handoff checkpoint deployment does not match the active deployment",
-            ));
-        }
-        if committed.scope != CheckpointScope::Cluster {
-            return Err(handoff_error(
-                "vnode handoff requires a cluster-scoped committed checkpoint",
-            ));
-        }
-        if !committed.reassignment_portable
-            || committed.assignment_fence.as_ref() != Some(predecessor)
-            || u32::from(committed.vnode_count) != predecessor.vnode_count
-        {
-            return Err(handoff_error(
-                "handoff checkpoint is not portable or does not cover the exact predecessor assignment",
-            ));
-        }
-
-        let mut requested_by_donor = BTreeMap::<u64, Vec<u16>>::new();
-        for &vnode in acquired_vnodes {
-            let vnode16 = u16::try_from(vnode)
-                .map_err(|_| handoff_error(format!("acquired vnode {vnode} exceeds u16")))?;
-            let donor = predecessor_owners[vnode as usize].0;
-            requested_by_donor.entry(donor).or_default().push(vnode16);
-        }
-        let mut expected_by_donor = requested_by_donor
-            .keys()
-            .copied()
-            .map(|donor| (donor, Vec::new()))
-            .collect::<BTreeMap<_, Vec<u16>>>();
-        for (vnode, owner) in predecessor_owners.iter().enumerate() {
-            if let Some(expected) = expected_by_donor.get_mut(&owner.0) {
-                expected.push(u16::try_from(vnode).map_err(|_| {
-                    handoff_error("predecessor vnode owner map exceeds the checkpoint ABI")
-                })?);
-            }
-        }
-
-        let donors = requested_by_donor
-            .into_iter()
-            .map(|(participant_id, requested)| {
-                let participant = committed
-                    .participants
-                    .binary_search_by_key(&participant_id, |entry| entry.participant_id)
-                    .ok()
-                    .map(|index| committed.participants[index].clone())
-                    .ok_or_else(|| {
-                        handoff_error(format!(
-                            "vnode donor {participant_id} is absent from the committed checkpoint"
-                        ))
-                    })?;
-                let expected = expected_by_donor.remove(&participant_id).ok_or_else(|| {
-                    handoff_error(format!(
-                        "vnode donor {participant_id} has no predecessor ownership roster"
-                    ))
-                })?;
-                Ok((participant, requested, expected))
-            })
-            .collect::<Result<Vec<_>, DbError>>()?;
-
-        let manifest_bytes = donors.iter().try_fold(0usize, |total, donor| {
-            let bytes = usize::try_from(donor.0.manifest_len).map_err(|_| {
-                DbError::ManagedStateBudgetExceeded {
-                    context: "[LDB-6050] vnode handoff manifests".into(),
-                    accounted_bytes: usize::MAX,
-                    limit_bytes: max_payload_bytes,
-                }
-            })?;
-            total
-                .checked_add(bytes)
-                .ok_or_else(|| DbError::ManagedStateBudgetExceeded {
-                    context: "[LDB-6050] vnode handoff manifests".into(),
-                    accounted_bytes: usize::MAX,
-                    limit_bytes: max_payload_bytes,
-                })
-        })?;
-        if manifest_bytes > max_payload_bytes {
-            return Err(DbError::ManagedStateBudgetExceeded {
-                context: "[LDB-6050] vnode handoff manifests".into(),
-                accounted_bytes: manifest_bytes,
-                limit_bytes: max_payload_bytes,
-            });
-        }
-
-        let checkpoint_id = committed.checkpoint_id;
-        let checkpoint_epoch = committed.epoch;
-        let committed_vnode_count = committed.vnode_count;
-        let manifest_reads = donors
-            .into_iter()
-            .map(|(participant, requested, expected)| {
-                let store = Arc::clone(&self.store);
-                let pipeline_identity = &pipeline_identity;
-                let deployment_id = &deployment_id;
-                async move {
-                    let participant_id = participant.participant_id;
-                    let manifest = tokio::time::timeout_at(
-                        deadline,
-                        store.load_manifest_verified(
-                            participant_id,
-                            checkpoint_id,
-                            participant.manifest_len,
-                            &participant.manifest_sha256,
-                        ),
-                    )
-                    .await
-                    .map_err(|_| {
-                        handoff_error(format!(
-                            "participant {participant_id} handoff manifest read timed out"
-                        ))
-                    })?
-                    .map_err(|error| {
-                        handoff_error(format!(
-                            "participant {participant_id} handoff manifest read failed: {error}"
-                        ))
-                    })?
-                    .ok_or_else(|| {
-                        handoff_error(format!(
-                            "participant {participant_id} handoff manifest is missing"
-                        ))
-                    })?;
-                    if manifest.participant_id != participant_id
-                        || manifest.node_data.chunk.participant_id != participant_id
-                        || manifest.node_data.object_length != participant.node_data_len
-                        || manifest.node_data.sha256 != participant.node_data_sha256
-                        || manifest.epoch != checkpoint_epoch
-                        || manifest.checkpoint_id != checkpoint_id
-                        || manifest.deployment_id != deployment_id.as_str()
-                        || &manifest.pipeline_identity != pipeline_identity
-                        || manifest.vnode_count != committed_vnode_count
-                        || manifest.assignment_fence.as_ref() != Some(predecessor)
-                        || manifest.owned_vnodes != expected
-                    {
-                        return Err(handoff_error(format!(
-                        "participant {participant_id} manifest does not match the exact handoff cut"
-                    )));
-                    }
-
-                    let selected = manifest
-                        .state_frames
-                        .iter()
-                        .filter(|frame| match &frame.key {
-                            StateFrameKey::OperatorWhole { operator_id } => {
-                                include_whole
-                                    && operator_id
-                                        .strip_prefix("graph:")
-                                        .is_some_and(|suffix| !suffix.is_empty())
-                            }
-                            StateFrameKey::Vnode { operator_id, vnode } => {
-                                operator_id
-                                    .strip_prefix("graph:")
-                                    .is_some_and(|suffix| !suffix.is_empty())
-                                    && requested.binary_search(vnode).is_ok()
-                            }
-                        })
-                        .cloned()
-                        .collect::<Vec<_>>();
-                    let selected_bytes = selected.iter().try_fold(0usize, |total, frame| {
-                        let bytes = usize::try_from(frame.range.length).map_err(|_| {
-                            DbError::ManagedStateBudgetExceeded {
-                                context: "[LDB-6050] vnode handoff payload".into(),
-                                accounted_bytes: usize::MAX,
-                                limit_bytes: max_payload_bytes,
-                            }
-                        })?;
-                        total.checked_add(bytes).ok_or_else(|| {
-                            DbError::ManagedStateBudgetExceeded {
-                                context: "[LDB-6050] vnode handoff payload".into(),
-                                accounted_bytes: usize::MAX,
-                                limit_bytes: max_payload_bytes,
-                            }
-                        })
-                    })?;
-                    let plan = VerifiedStateFramePlan::new(&manifest, &selected)?;
-                    Ok((plan, selected_bytes))
-                }
-            });
-        let loaded = futures::stream::iter(manifest_reads)
-            .buffer_unordered(MAX_RETENTION_IO_CONCURRENCY)
-            .try_collect::<Vec<(VerifiedStateFramePlan, usize)>>()
-            .await?;
-
-        let payload_bytes = loaded.iter().try_fold(0usize, |total, (_, bytes)| {
-            total
-                .checked_add(*bytes)
-                .ok_or_else(|| DbError::ManagedStateBudgetExceeded {
-                    context: "[LDB-6050] vnode handoff payload".into(),
-                    accounted_bytes: usize::MAX,
-                    limit_bytes: max_payload_bytes,
-                })
-        })?;
-        if payload_bytes > max_payload_bytes {
-            return Err(DbError::ManagedStateBudgetExceeded {
-                context: "[LDB-6050] vnode handoff payload".into(),
-                accounted_bytes: payload_bytes,
-                limit_bytes: max_payload_bytes,
-            });
-        }
-        if tokio::time::Instant::now() >= deadline {
-            return Err(handoff_error("vnode handoff read timed out"));
-        }
-
-        let plans = loaded.into_iter().map(|(plan, _)| plan).collect();
-        tokio::time::timeout_at(
-            deadline,
-            load_verified_state_frames(self.store.as_ref(), plans),
-        )
-        .await
-        .map_err(|_| handoff_error("vnode handoff frame read timed out"))?
-    }
-
-    fn recovery_scope(&self) -> CheckpointScope {
-        #[cfg(feature = "cluster")]
-        if self.cluster_controller.is_some() {
-            return CheckpointScope::Cluster;
-        }
-        CheckpointScope::Local
-    }
-
-    fn manifest_watermark(
-        manifest: Option<&CheckpointManifest>,
-    ) -> Result<laminar_core::checkpoint::CheckpointWatermark, DbError> {
-        let Some(manifest) = manifest else {
-            return Ok(laminar_core::checkpoint::CheckpointWatermark::Uninitialized);
-        };
-        classify_channel_progress(&manifest.channel_progress)
-            .map_err(|error| DbError::Checkpoint(format!("recovered channel progress: {error}")))
-    }
-
-    async fn install_recovered_cut(
-        &mut self,
-        outcome: laminar_core::checkpoint_decision::CheckpointOutcome,
-        committed: CommittedCheckpointIndex,
-        deadline: tokio::time::Instant,
-    ) -> Result<crate::recovery_manager::RecoveredState, DbError> {
-        let pipeline_identity = self.expected_pipeline_identity()?;
-        let deployment_id = self.expected_deployment_id()?.to_owned();
-        let recovery_scope = self.recovery_scope();
-        #[cfg(feature = "cluster")]
-        let cluster_target = if recovery_scope == CheckpointScope::Cluster {
-            let controller = self.cluster_controller.as_ref().ok_or_else(|| {
-                DbError::Checkpoint("cluster recovery has no cluster controller".into())
-            })?;
-            let assignment = controller
-                .checkpoint_assignment_fence(self.assignment_version)
-                .ok_or_else(|| {
-                    DbError::Checkpoint(format!(
-                        "cluster recovery has no active assignment fence for version {}",
-                        self.assignment_version
-                    ))
-                })?;
-            Some(crate::recovery_manager::ClusterRecoveryTarget {
-                assignment,
-                owned_vnodes: self.owned_vnodes.clone(),
-                max_graph_payload_bytes: self.recovery_graph_payload_limit,
-            })
-        } else {
-            None
-        };
-        #[cfg(not(feature = "cluster"))]
-        let cluster_target = None;
-        #[cfg(feature = "cluster")]
-        if let Some(target) = cluster_target.as_ref() {
-            let predecessor = committed.assignment_fence.as_ref().ok_or_else(|| {
-                DbError::Checkpoint("cluster recovery checkpoint has no assignment fence".into())
-            })?;
-            if predecessor != &target.assignment {
-                if predecessor.assignment_version >= target.assignment.assignment_version
-                    || predecessor.vnode_count != target.assignment.vnode_count
-                    || predecessor.partitioning_abi_version
-                        != target.assignment.partitioning_abi_version
-                    || !committed.reassignment_portable
-                {
-                    return Err(DbError::Checkpoint(format!(
-                        "recovery target assignment {} is not a compatible newer bootstrap target for committed assignment {}",
-                        target.assignment.assignment_version, predecessor.assignment_version
-                    )));
-                }
-                let expected = outcome.committed_checkpoint.as_ref().ok_or_else(|| {
-                    DbError::Checkpoint(
-                        "cluster Commit outcome has no committed checkpoint reference".into(),
-                    )
-                })?;
-                let controller = self.cluster_controller.as_ref().ok_or_else(|| {
-                    DbError::Checkpoint("cluster recovery has no cluster controller".into())
-                })?;
-                let authority = controller.checkpoint_authority().map_err(|error| {
-                    DbError::Checkpoint(format!("cluster checkpoint authority: {error}"))
-                })?;
-                let pinned = tokio::time::timeout_at(
-                    deadline,
-                    authority.assignment_handoff_checkpoint(&target.assignment),
-                )
-                .await
-                .map_err(|_| {
-                    DbError::Checkpoint("assignment handoff checkpoint lookup timed out".into())
-                })?
-                .map_err(|error| {
-                    DbError::Checkpoint(format!(
-                        "assignment handoff checkpoint lookup failed: {error}"
-                    ))
-                })?;
-                if pinned.as_ref() != Some(expected) {
-                    return Err(DbError::Checkpoint(
-                        "recovery checkpoint is not the durable handoff pin for the active assignment"
-                            .into(),
-                    ));
-                }
-            }
-        }
-        let recovered = {
-            let manager = crate::recovery_manager::RecoveryManager::new(
-                self.store.as_ref(),
-                &pipeline_identity,
-                &deployment_id,
-                recovery_scope,
-            );
-            tokio::time::timeout_at(
-                deadline,
-                manager.recover_committed_for_target(&outcome, &committed, cluster_target),
-            )
-            .await
-            .map_err(|_| DbError::Checkpoint("checkpoint recovery timed out".into()))??
-        };
-
-        #[cfg(feature = "cluster")]
-        let continuation_proof = if committed.scope == CheckpointScope::Cluster {
-            self.cluster_controller
-                .as_ref()
-                .and_then(|controller| controller.capture_leader_proof())
-        } else {
-            None
-        };
-        #[cfg(not(feature = "cluster"))]
-        let continuation_proof: Option<LeaderProof> = None;
-        let continuation_fencing_token = match committed.scope {
-            CheckpointScope::Local => Some(1),
-            CheckpointScope::Cluster => {
-                #[cfg(feature = "cluster")]
-                {
-                    recovery_sink_fence(outcome.leader_proof.as_ref(), continuation_proof.as_ref())?
-                }
-                #[cfg(not(feature = "cluster"))]
-                {
-                    None
-                }
-            }
-        };
-        if let Some(fencing_token) = continuation_fencing_token {
-            let manifests = recovered.manifests.iter().collect::<Vec<_>>();
-            self.commit_external_sinks_until(
-                CheckpointAttempt::canonical(committed.checkpoint_id),
-                &manifests,
-                fencing_token,
-                committed
-                    .predecessor
-                    .as_ref()
-                    .map_or(0, |reference| reference.checkpoint_id),
-                deadline,
-            )
-            .await?;
-            self.schedule_retention(committed.clone(), continuation_proof.as_ref());
-        }
-
-        let reference = outcome.committed_checkpoint.clone().ok_or_else(|| {
-            DbError::Checkpoint("Commit outcome has no committed checkpoint reference".into())
-        })?;
-        let committed_source_watermarks = committed
-            .effective_source_watermarks()
-            .map_err(DbError::Checkpoint)?;
-        let local_manifest = (!recovered.reassigned)
-            .then(|| {
-                recovered
-                    .manifests
-                    .iter()
-                    .find(|manifest| manifest.participant_id == self.store.participant_id())
-                    .cloned()
-                    .map(Arc::new)
-            })
-            .flatten();
-        self.local_watermark = Self::manifest_watermark(local_manifest.as_deref())?;
-        self.last_committed_manifest = local_manifest;
-        self.last_committed_ref = Some(reference.clone());
-        self.last_committed_source_watermarks =
-            Some((reference, committed_source_watermarks.clone()));
-        self.prepared.clear();
-        self.allocator.advance_epoch_to(checked_successor_epoch(
-            committed.epoch,
-            "installing recovered checkpoint",
-        )?);
-        self.failure_requires_recovery = false;
-        self.phase = CheckpointPhase::Idle;
-        #[cfg(feature = "cluster")]
-        if let Some(controller) = self.cluster_controller.as_ref() {
-            controller
-                .replace_recovered_checkpoint_progress(
-                    &committed.channel_progress,
-                    &committed_source_watermarks,
-                )
-                .map_err(DbError::Checkpoint)?;
-        }
-        Ok(recovered)
-    }
-
-    pub async fn recover(
-        &mut self,
-    ) -> Result<Option<crate::recovery_manager::RecoveredState>, DbError> {
-        if self.phase != CheckpointPhase::Idle {
-            return Err(DbError::Checkpoint(
-                "cannot recover while a checkpoint is in progress".into(),
-            ));
-        }
-        let deadline = tokio::time::Instant::now() + self.config.checkpoint_timeout;
-
-        #[cfg(feature = "cluster")]
-        if let Some(controller) = self.cluster_controller.as_ref() {
-            let authority = controller.checkpoint_authority().map_err(|error| {
-                DbError::Checkpoint(format!("cluster checkpoint authority: {error}"))
-            })?;
-            let Some(selected) =
-                tokio::time::timeout_at(deadline, authority.highest_cluster_committed_outcome())
-                    .await
-                    .map_err(|_| {
-                        DbError::Checkpoint("cluster recovery selection timed out".into())
-                    })?
-                    .map_err(|error| {
-                        DbError::Checkpoint(format!("cluster recovery selection: {error}"))
-                    })?
-            else {
-                return Ok(None);
-            };
-            let (outcome, committed) = tokio::time::timeout_at(
-                deadline,
-                authority.cluster_outcome_with_committed_checkpoint(selected.epoch),
-            )
-            .await
-            .map_err(|_| DbError::Checkpoint("cluster checkpoint read timed out".into()))?
-            .map_err(|error| DbError::Checkpoint(format!("cluster checkpoint read: {error}")))?
-            .ok_or_else(|| DbError::Checkpoint("selected cluster checkpoint disappeared".into()))?;
-            if outcome != selected {
-                return Err(DbError::Checkpoint(
-                    "cluster recovery selection changed during exact read".into(),
-                ));
-            }
-            let committed = committed.ok_or_else(|| {
-                DbError::Checkpoint("selected cluster Commit has no checkpoint index".into())
-            })?;
-            return self
-                .install_recovered_cut(outcome, committed, deadline)
-                .await
-                .map(Some);
-        }
-
-        let store = Arc::clone(self.decision_store.as_ref().ok_or_else(|| {
-            DbError::Checkpoint("checkpoint recovery requires a decision store".into())
-        })?);
-        let mut head = tokio::time::timeout_at(deadline, store.checkpoint_decision_head())
-            .await
-            .map_err(|_| DbError::Checkpoint("checkpoint recovery selection timed out".into()))?
-            .map_err(|error| {
-                DbError::Checkpoint(format!("checkpoint recovery selection: {error}"))
-            })?;
-        if let Some(inventory) = head.as_ref().and_then(|head| head.active_artifacts.clone()) {
-            self.validate_checkpoint_artifact_inventory(&inventory)?;
-            if inventory.assignment_fence.is_some() {
-                return Err(DbError::Checkpoint(
-                    "local recovery found cluster artifact inventory".into(),
-                ));
-            }
-            let exact_abort = head
-                .as_ref()
-                .and_then(|head| head.latest_terminal.as_ref())
-                .is_some_and(|outcome| {
-                    !outcome.is_commit()
-                        && outcome.epoch == inventory.attempt.epoch
-                        && outcome.checkpoint_id == inventory.attempt.checkpoint_id
-                });
-            if !exact_abort {
-                if head
-                    .as_ref()
-                    .and_then(|head| head.latest_terminal.as_ref())
-                    .is_some_and(|outcome| outcome.epoch >= inventory.attempt.epoch)
-                {
-                    return Err(DbError::Checkpoint(format!(
-                        "checkpoint {} has incompatible terminal authority while artifacts remain",
-                        inventory.attempt.checkpoint_id
-                    )));
-                }
-                self.record_outcome_until(
-                    inventory.attempt,
-                    laminar_core::checkpoint_decision::CheckpointVerdict::Abort,
-                    None,
-                    None,
-                    None,
-                    deadline,
-                )
-                .await?;
-            }
-            self.cleanup_local_checkpoint_artifacts_until(inventory.attempt, deadline)
-                .await?;
-            head = tokio::time::timeout_at(deadline, store.checkpoint_decision_head())
-                .await
-                .map_err(|_| DbError::Checkpoint("checkpoint recovery selection timed out".into()))?
-                .map_err(|error| {
-                    DbError::Checkpoint(format!("checkpoint recovery selection: {error}"))
-                })?;
-        }
-        let Some(outcome) = head.and_then(|head| head.latest_commit) else {
-            return Ok(None);
-        };
-        let reference = outcome.committed_checkpoint.as_ref().ok_or_else(|| {
-            DbError::Checkpoint("selected Commit has no checkpoint index reference".into())
-        })?;
-        let committed =
-            tokio::time::timeout_at(deadline, store.load_committed_checkpoint(reference))
-                .await
-                .map_err(|_| DbError::Checkpoint("committed checkpoint read timed out".into()))?
-                .map_err(|error| {
-                    DbError::Checkpoint(format!("committed checkpoint read: {error}"))
-                })?;
-        self.install_recovered_cut(outcome, committed, deadline)
-            .await
-            .map(Some)
-    }
-
-    #[cfg(feature = "cluster")]
-    pub async fn recover_to_epoch(
-        &mut self,
-        epoch: u64,
-    ) -> Result<Option<crate::recovery_manager::RecoveredState>, DbError> {
-        if self.phase != CheckpointPhase::Idle {
-            return Err(DbError::Checkpoint(
-                "cannot recover while a checkpoint is in progress".into(),
-            ));
-        }
-        let deadline = tokio::time::Instant::now() + self.config.checkpoint_timeout;
-        if epoch == 0 {
-            let controller = self.cluster_controller.as_ref().ok_or_else(|| {
-                DbError::Checkpoint(
-                    "epoch-targeted recovery requires cluster checkpoint authority".into(),
-                )
-            })?;
-            let authority = controller.checkpoint_authority().map_err(|error| {
-                DbError::Checkpoint(format!("cluster checkpoint authority: {error}"))
-            })?;
-            if tokio::time::timeout_at(deadline, authority.cluster_checkpoint_artifacts())
-                .await
-                .map_err(|_| {
-                    DbError::Checkpoint("genesis recovery artifact audit timed out".into())
-                })?
-                .map_err(|error| {
-                    DbError::Checkpoint(format!("genesis recovery artifact audit failed: {error}"))
-                })?
-                .is_some()
-            {
-                return Err(DbError::Checkpoint(
-                    "genesis recovery cannot install while checkpoint artifacts remain".into(),
-                ));
-            }
-            if let Some(reference) = self
-                .authoritative_committed_predecessor_until(CheckpointScope::Cluster, deadline)
-                .await?
-            {
-                return Err(DbError::Checkpoint(format!(
-                    "genesis recovery cannot replace authoritative committed checkpoint {}",
-                    reference.checkpoint_id
-                )));
-            }
-            self.prepared.clear();
-            self.last_committed_manifest = None;
-            self.last_committed_ref = None;
-            self.last_committed_source_watermarks = None;
-            self.local_watermark = CheckpointWatermark::Uninitialized;
-            self.failure_requires_recovery = false;
-            self.phase = CheckpointPhase::Idle;
-            controller
-                .replace_recovered_checkpoint_progress(&[], &BTreeMap::new())
-                .map_err(DbError::Checkpoint)?;
-            return Ok(None);
-        }
-        let authoritative = self
-            .authoritative_committed_predecessor_until(CheckpointScope::Cluster, deadline)
-            .await?;
-        let authoritative = authoritative.ok_or_else(|| {
-            DbError::Checkpoint(format!(
-                "cluster epoch {epoch} cannot be recovered because no authoritative Commit exists"
-            ))
-        })?;
-        if authoritative.epoch != epoch || authoritative.checkpoint_id != epoch {
-            return Err(DbError::Checkpoint(format!(
-                "cluster epoch {epoch} is not the authoritative committed recovery head {}",
-                authoritative.checkpoint_id
-            )));
-        }
-        if let Some(controller) = self.cluster_controller.as_ref() {
-            let authority = controller.checkpoint_authority().map_err(|error| {
-                DbError::Checkpoint(format!("cluster checkpoint authority: {error}"))
-            })?;
-            let (outcome, committed) = tokio::time::timeout_at(
-                deadline,
-                authority.cluster_outcome_with_committed_checkpoint(epoch),
-            )
-            .await
-            .map_err(|_| DbError::Checkpoint("cluster checkpoint read timed out".into()))?
-            .map_err(|error| DbError::Checkpoint(format!("cluster checkpoint read: {error}")))?
-            .ok_or_else(|| DbError::Checkpoint(format!("cluster epoch {epoch} has no outcome")))?;
-            let committed = committed.ok_or_else(|| {
-                DbError::Checkpoint(format!("cluster epoch {epoch} is not committed"))
-            })?;
-            if outcome.committed_checkpoint.as_ref() != Some(&authoritative) {
-                return Err(DbError::Checkpoint(
-                    "cluster recovery outcome does not match the authoritative committed head"
-                        .into(),
-                ));
-            }
-            return self
-                .install_recovered_cut(outcome, committed, deadline)
-                .await
-                .map(Some);
-        }
-
-        Err(DbError::Checkpoint(
-            "epoch-targeted recovery requires cluster checkpoint authority".into(),
-        ))
     }
 
     async fn create_sink_witness_until(
